Fixed EDL when it is a single movie file. Added gray backgrounds to a number of requesters and windows. Fixed hotkeys for view, display and input color space. Made progress report be readable on black color theme. Made file chooser use idle callbacks instead of timeouts or threads.
Former-commit-id: 1f9a401a5ae67973a9c9fafef0d501dd77cdce81
Former-commit-id: 9821b8b4c4e052aec8d434719b72434e8aea8a87

+++ mrViewer/src/video/mrvGLEngine.cpp
@@ -3905,7 +3905,7 @@ void GLEngine::loadOpenGLShader()
         avcol_pri_to_mp_csp_prim( c->color_primaries ), // primaries
         avcol_trc_to_mp_csp_trc( c->color_trc ),   // gamma
         MP_CSP_LIGHT_DISPLAY,  // light
-        max_cll / MP_REF_WHITE               // sig_peak
+        (float)max_cll / (float)MP_REF_WHITE               // sig_peak
     };
 
     mp_colorspace dst
